1. Got rid of any states in fragment
