+++ app/src/main/java/com/example/playlistmaker/library/ui/model/FavoritesState.kt
@@ -4,10 +4,11 @@ import com.example.playlistmaker.search.domain.model.Track
 
 sealed interface FavoritesState {
     object Loading : FavoritesState
-    class Error(val error: FavoritesError) : FavoritesState
+    data class Error(val error: FavoritesError) : FavoritesState
     sealed class Success(open val tracks: List<Track>) : FavoritesState {
-        class FavoriteContent(
+        data class FavoriteContent(
             override val tracks: List<Track>
         ) : Success(tracks)
     }
+    data class NavigateToPlayer(val track: Track):FavoritesState
 }
